fix editor.css generation with MODULARITY_COMPILE

+++ includes/Modularity.php
@@ -89,7 +89,7 @@ if (!class_exists("Modularity")) {
     }
 
     private function generateEditorCSS($compiler, $scssFile, $scssCode) {
-      if (!strpos($scssCode, "generate_editor_styles")) {
+      if (!defined('MODULARITY_COMPILE') && !strpos($scssCode, "generate_editor_styles")) {
         $cssFileEditor = str_replace(".scss", ".editor.css", $scssFile);
         if (file_exists($cssFileEditor)) unlink($cssFileEditor);
         return;
